requirements.txt
scikit-learn
matplotlib
transformers
accelerate
peft
evaluate
tqdm
datasets
Pillow
torchvision
huggingface_hub
safetensors
wandb
googlesearch-python
diffusers=0.11.1
attrs
platformdirs
omegaconf
pycocotools
